Add Day 6 Challenges

+++ src/main/kotlin/Days/Day.kt
@@ -2,6 +2,6 @@ package Days
 
 interface Day {
     val DAY: Int
-    fun challenge1(): Int
-    fun challenge2(): Int
+    fun challenge1(): Long
+    fun challenge2(): Long
 }
